consolidate functions in one file
