--- src/features/movies/getPopularMovies.js
@@ -1,11 +1,0 @@
-export const getPopularMovies = async (page) => {
-    const apiURL = `https://api.themoviedb.org/3/movie/popular?api_key=768f7875782193f5e4797762314da0b7&page=${page}&language=en-US`;
-
-    const response = await fetch(apiURL);
-
-    if (!response.ok) {
-        new Error(response.statusText);
-    }
-
-    return await response.json();
-};
